+++ app/src/main/java/com/tonykuz/funkyfood/domain/Film.kt
@@ -1,4 +1,4 @@
-package com.tonykuz.funkyfood
+package com.tonykuz.funkyfood.domain
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
@@ -7,6 +7,6 @@ import kotlinx.parcelize.Parcelize
 data class Film(
     val title: String,
     val poster: Int,
-    val description: String,
+    val instructions: String,
     var isInFavorites: Boolean = false
 ) : Parcelable
