+++ app/src/main/res/drawable/background_search_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="@android:color/darker_gray"/>
+    <solid android:color="@color/white"/>
     <corners android:radius="15dp"/>
 </shape>
