+++ app/src/main/res/layout/film_item.xml
@@ -43,7 +43,7 @@
             app:layout_constraintLeft_toRightOf="@+id/poster"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/title"
-            tools:text="@tools:sample/lorem/random" />
+            tools:text="Description" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.cardview.widget.CardView>
