+++ app/src/main/res/menu/menu_navi.xml
@@ -2,6 +2,11 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
+        android:id="@+id/home"
+        android:icon="@drawable/ic_round_home"
+        android:title="@string/menu_home_title_text"
+        app:showAsAction="withText"/>
+    <item
         android:id="@+id/favorites"
         android:icon="@drawable/ic_round_favorite"
         android:title="@string/menu_favorites_title"
