--- app/src/main/res/menu/toolbar_menu.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/settings"
-        android:title="@string/menu_settings_title"
-        android:icon="@drawable/ic_round_settings"
-        app:showAsAction="ifRoom"/>
-</menu>
